feat: restructure RA.Aid scripts and improve session usage statistics retrieval
- Move session usage functionality to a dedicated module for better organization.
- Create a command-line interface for RA.Aid scripts to enhance usability.
- Update Makefile to reflect changes in script paths.
- Add a new script to extract changelog entries for specific versions.
- Implement a script to generate predictions for SWE-bench Lite, improving functionality.
- Remove deprecated code and streamline imports for better maintainability.
