+++ pyproject.toml
@@ -77,5 +77,5 @@ version = {attr = "ra_aid.version.__version__"}
 path = "ra_aid/__version__.py"
 
 [tool.hatch.build.targets.wheel]
-packages = ["ra_aid", "scripts"]
+packages = ["ra_aid"]
