+++ ra_aid/scripts/__init__.py
@@ -4,6 +4,6 @@ RA.Aid utility scripts.
 This package contains utility scripts for RA.Aid.
 """
 
-from .get_session_usage import get_latest_session_usage, create_empty_result
+from .session_usage import get_latest_session_usage, create_empty_result
 
 __all__ = ['get_latest_session_usage', 'create_empty_result']
